AllPostsBasePage was created, AllPostsPage and TrashPostsPage were refactored due to this
other classes were cleaned
