+++ src/test/java/ui/model/posts/AllPostsPage.java
@@ -2,46 +2,13 @@ package ui.model.posts;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
-import org.openqa.selenium.support.FindBy;
-import ui.model.base.BasePage;
-import ui.model.base.Header;
-import ui.model.base.SidePanel;
+import ui.model.base.AllPostsBasePage;
 
-import java.util.List;
-
-public class AllPostsPage extends BasePage {
-
-    private Header header;
-
-    private SidePanel sidePanel;
-
-    @FindBy(css = "#the-list > tr > td > strong")
-    private List<WebElement> postsList;
-
-    @FindBy(id = "cb-select-all-1")
-    private WebElement allPostsCheckbox;
-
-    @FindBy(id = "cb-select-1")
-    private WebElement helloWorldPostCheckbox;
-
-    @FindBy(id = "bulk-action-selector-top")
-    private WebElement bulkActionSelector;
-
-    @FindBy(css = "option[value=\"trash\"]")
-    private WebElement moveToTrashOption;
-
-    @FindBy(id = "doaction")
-    private WebElement applyButton;
-
-    @FindBy(xpath = "//a[@href='edit.php?post_status=trash&post_type=post']")
-    private WebElement trashLink;
+public class AllPostsPage extends AllPostsBasePage {
 
     public AllPostsPage(WebDriver driver) {
         super(driver);
-        header = new Header(driver);
-        sidePanel = new SidePanel(driver);
     }
 
     public AllPostsPage hoverOnPostTitle(String titleName) {
@@ -53,53 +20,11 @@ public class AllPostsPage extends BasePage {
         return this;
     }
 
-    public AllPostsPage clickTrashButton(String titleName) {
+    public AllPostsPage clickTrashButtonUnderPostTitle(String titleName) {
         getDriver()
                 .findElement(By.xpath(String.format("//span[@class='trash']/a[contains(@aria-label, '%s')]", titleName)))
                 .click();
 
         return new AllPostsPage(getDriver());
-    }
-
-    public int getPostsListSize() {
-        return postsList.size();
-    }
-
-    public Header getHeader() {
-        return header;
-    }
-
-    public SidePanel getSidePanel() {
-        return sidePanel;
-    }
-
-    public AllPostsPage clickSelectAllPostsCheckbox() {
-        allPostsCheckbox.click();
-        return this;
-    }
-
-    public AllPostsPage clickHelloWorldPostCheckbox() {
-        helloWorldPostCheckbox.click();
-        return this;
-    }
-
-    public AllPostsPage clickBulkActionSelector() {
-        bulkActionSelector.click();
-        return this;
-    }
-
-    public AllPostsPage clickMoveToTrashOption() {
-        moveToTrashOption.click();
-        return this;
-    }
-
-    public AllPostsPage clickApplyButton() {
-        applyButton.click();
-        return this;
-    }
-
-    public TrashPostsPage clickTrashLink() {
-        trashLink.click();
-        return new TrashPostsPage(getDriver());
     }
 }
